Correct getTypeByKey()  for optional properties (#121)
When property is optional then returned type must also allow undefined. So if not already a union type containing undefined a corresponding union type is now created and returned.

Also had to fix result of type-indexed-access-object-2 test. The output doesn't make much sense because it isn't possible to use undefined in JSON anyway but then the test itself may not make any sense because it actually uses an optional value as JSON root type.