Improve condition type node parser
When check-type is a type parameter then we actually have to follow both branches with the check-type narrowed down by the extends-type (And the inverse of it) and then create a union type of all results. At least I hope that this finally gets it right now...

To prevent useless duplications in the resulting type I also improved the union type itself so it automatically flattens itself if possible, removes duplicate types and Never types. The new normalize() method can be used to convert it to the single type when only one is present or Never type if union is empty.