Make more of createParser’s config optional
This is another PR along the lines of #385.

The CLI has a few defaults which the programmatic interface doesn’t. I think it makes sense to harmonize these. In my use case, I’m using the CLi to generate the typings, and a test to check that they’re up to date. It would be better not to have to pass in these arguments (which I wasn’t even aware of when I was using the CLI).

Another way to implement this would be to create a MergedConfig type; happy to go that route if you’d prefer.

Looking forward to your feedback!