Better output for idiom [A, A, ...A[]]
The type expression above can be used to indicate a uniform array with
a minimum number of elements.

Prior to this change, output for the expression looks roughly as
follows (schema references omitted for clarity):

    {
      "type": "array",
      "items": [
        {
          "type": "A"
        },
        {
          "type": "A"
        }
      ],
      "minItems": 2,
      "additionalItems": {
        "type": "A"
      }
    }

While this is technically correct, and a direct translation of the
input, it obscures intent.  It also causes a warning to be emitted in
Ajv 7.x, see https://github.com/ajv-validator/ajv/issues/1417

With this change the output looks as below and doesn't generate a
warning:

    {
      "type": "array",
      "items": {
        "type": "A"
      },
      "minItems": 2,
    }