Adding functional tests, mergeDefinitions tests
In addition to the raw schema tests which just check that nothing has
changed since the last snapshot, this adds makeExemplar, which builds
a raw value using the parsed BaseType representation of a TypeScript
type tree. This way, tests can specify valid/invalid data to be verified
using the Ajv validator in addition to raw schema output tests.