Handle intersections of aliased unions
ts-json-schema-generator correctly handles basic types like `(A | B) & C`, but if A is itself an alias for `A1 | A2`, it fails.

Delete the FIXME comment and the reference to #62; if I understand correctly, this is correctly handled as of d9b9a15aa5849a2e298c009f60ada0d3539b6f47.